safety?

+++ src/thing_stats.c
@@ -911,7 +911,7 @@ TbBool update_relative_creature_health(struct Thing* creatng)
     JUSTLOG("old_max_health: %d", (int)cctrl->max_health);
     cctrl->max_health = calculate_correct_creature_max_health(creatng);
     JUSTLOG("new_max_health: %d", (int)cctrl->max_health);
-    int64_t health_scaled = (int64_t)cctrl->max_health * health_permil / 1000;
+    int64_t health_scaled = (int64_t)cctrl->max_health * (int64_t)health_permil / 1000;
     JUSTLOG("health_scaled: %lld", health_scaled);
     creatng->health = health_scaled;
     JUSTLOG("new_health: %d", (int)creatng->health);
